+++ cookiecutter/{{cookiecutter.organization}}-richie-site-factory/template/{{cookiecutter.site}}/src/backend/{{cookiecutter.site}}/settings.py
@@ -661,10 +661,10 @@ class Base(StyleguideMixin, DRFMixin, RichieCoursesConfigurationMixin, Configura
 
     # Wheither you can create PageIndex extension on page through toolbar if true or
     # just editing existing extension if false
-    RICHIE_PAGEINDEX_ALLOW_CREATION = False
+    RICHIE_MAINMENUENTRY_ALLOW_CREATION = False
 
     # Define which node level can be processed to search for pageindex extension
-    RICHIE_PAGEINDEX_MENU_ALLOWED_LEVEL = 0
+    RICHIE_MAINMENUENTRY_MENU_ALLOWED_LEVEL = 0
 
     # pylint: disable=invalid-name
     @property