+++ src/richie/apps/courses/admin.py
@@ -280,9 +280,9 @@ class CourseAdmin(FrontendEditableAdminMixin, PageExtensionAdmin):
         return JsonResponse({"id": new_page.course.id})
 
 
-class IndexPageAdmin(PageExtensionAdmin):
+class MainMenuEntryAdmin(PageExtensionAdmin):
     """
-    Admin class for the IndexPage model
+    Admin class for the MainMenuEntry model
     """
 
     list_display = ["title", "allow_submenu"]
@@ -364,7 +364,7 @@ admin.site.register(models.Category, CategoryAdmin)
 admin.site.register(models.Course, CourseAdmin)
 admin.site.register(models.CourseRun, CourseRunAdmin)
 admin.site.register(models.Licence, LicenceAdmin)
-admin.site.register(models.IndexPage, IndexPageAdmin)
+admin.site.register(models.MainMenuEntry, MainMenuEntryAdmin)
 admin.site.register(models.Organization, OrganizationAdmin)
 admin.site.register(models.PageRole, PageRoleAdmin)
 admin.site.register(models.Person, PersonAdmin)