+++ src/richie/apps/courses/cms_toolbars.py
@@ -13,7 +13,7 @@ from cms.utils.page_permissions import user_can_add_subpage, user_can_change_pag
 from cms.utils.urlutils import admin_reverse
 
 from .defaults import PAGE_EXTENSION_TOOLBAR_ITEM_POSITION
-from .models import Category, Course, IndexPage, Organization, Person
+from .models import Category, Course, MainMenuEntry, Organization, Person
 
 
 class BaseExtensionToolbar(ExtensionToolbar):
@@ -135,21 +135,21 @@ class PersonExtensionToolbar(BaseExtensionToolbar):
 
 
 @toolbar_pool.register
-class IndexPageExtensionToolbar(BaseExtensionToolbar):
+class MainMenuEntryExtensionToolbar(BaseExtensionToolbar):
     """
-    This extension class customizes the toolbar for the IndexPage page extension.
+    This extension class customizes the toolbar for the MainMenuEntry page extension.
     """
 
-    model = IndexPage
+    model = MainMenuEntry
 
     def populate(self):
         """
         Specific extension populate method.
 
         This extension entry only appears in toolbar if page already have extension or
-        if setting ``RICHIE_PAGEINDEX_ALLOW_CREATION`` is true. Finally the page level
-        must also match the allowed level from setting
-        ``RICHIE_PAGEINDEX_MENU_ALLOWED_LEVEL``.
+        if setting ``RICHIE_MAINMENUENTRY_ALLOW_CREATION`` is true. Finally the page
+        level must also match the allowed level from setting
+        ``RICHIE_MAINMENUENTRY_MENU_ALLOWED_LEVEL``.
         """
         # always use draft if we have a page
         self.page = get_page_draft(self.request.current_page)
@@ -167,16 +167,16 @@ class IndexPageExtensionToolbar(BaseExtensionToolbar):
             level = self.page.node.get_depth() - 1
             allowed = page_extension is not None or (
                 page_extension is None
-                and settings.RICHIE_PAGEINDEX_ALLOW_CREATION is True
+                and settings.RICHIE_MAINMENUENTRY_ALLOW_CREATION is True
             )
             if (
                 allowed
-                and level == settings.RICHIE_PAGEINDEX_MENU_ALLOWED_LEVEL
+                and level == settings.RICHIE_MAINMENUENTRY_MENU_ALLOWED_LEVEL
                 and admin_url
             ):
                 # Adds a toolbar item in position 0 (at the top of the menu)
                 page_menu.add_modal_item(
-                    _("Index Page settings"),
+                    _("Main menu settings"),
                     url=admin_url,
                     disabled=not self.toolbar.edit_mode_active,
                     position=PAGE_EXTENSION_TOOLBAR_ITEM_POSITION,